Lazify `dask_utils`
